Add makefile and modify test setting. Fix a bug in targets swap.
