+++ skdpu/tree/src/dpu_programs/test/test5.h
@@ -1,0 +1,115 @@
+/**
+ * @file test5.h
+ * @author Julien Legriel (jlegriel@upmem.com)
+ * @brief Test 5 for the DPU side of the tree algorithm
+ *
+ */
+
+#ifndef _TREES_DPU_KERNEL_TEST5_H_
+#define _TREES_DPU_KERNEL_TEST5_H_
+
+#include "./test.h"
+
+static void test_init() {
+
+  printf("--------------- TEST5 ----------------------\n");
+
+  // set points/classes
+  if (me() == 0) {
+    int k = 0;
+    n_points = 200;
+    n_features = 5;
+    for (int j = 0; j < n_features; ++j) {
+      for (int i = 0; i < n_points; ++i) {
+        if (i < 30 || (i >= 100 && i < 130))
+          t_features[k] = (j & 1) ? ((i != 5 && i != 9) ? 10 : 50) : 50;
+        else
+          t_features[k] = (j & 1) ? 50 : 10;
+        k++;
+      }
+    }
+    n_classes = 4;
+    for (int i = 0; i < n_points; ++i) {
+      t_targets[i] = i & 3;
+      printf("point %u target %f\n", i, t_targets[i]);
+    }
+    nb_cmds = 2;
+    struct Command cmd2, cmd3;
+
+    cmd2.type = 0;
+    cmd2.feature_index = 2;
+    cmd2.feature_threshold = 30;
+    cmd2.leaf_index = 0;
+
+    cmd3.type = 1;
+    cmd3.feature_index = 3;
+    cmd3.feature_threshold = 30;
+    cmd3.leaf_index = 1;
+
+    cmds_array[0] = cmd2;
+    cmds_array[1] = cmd3;
+
+    n_leaves = 2;
+    leaf_start_index[0] = 0;
+    // the tree leaf start of the split commit is going to be disaligned
+    // This forces the prolog case
+    leaf_end_index[0] = (n_points >> 1) - 1;
+    leaf_start_index[1] = (n_points >> 1) - 1;
+    leaf_end_index[1] = n_points;
+
+    for (int l = 0; l < n_leaves; ++l) {
+      printf("leaf %u:\n", l);
+      for (uint32_t f = 0; f < n_features; ++f) {
+        for (int i = leaf_start_index[l]; i < leaf_end_index[l]; ++i) {
+          printf("feature %u: %f\n", f, t_features[f * n_points + i]);
+        }
+      }
+    }
+  }
+}
+
+static void test_check() {
+
+  assert(n_leaves == 3);
+  uint32_t split_feature = 3;
+  uint32_t *expected_gini_cnt =
+      mem_alloc(n_leaves * n_classes * sizeof(uint32_t));
+  uint32_t *expected_leaf_cnt = mem_alloc(n_leaves * sizeof(uint32_t));
+  float *expected_feature_values =
+      mem_alloc(n_features * n_points * sizeof(float));
+  expected_gini_cnt[0] = 17;
+  expected_gini_cnt[1] = 17;
+  expected_gini_cnt[2] = 18;
+  expected_gini_cnt[3] = 17;
+  for (int i = 0; i < n_classes; ++i) {
+    expected_gini_cnt[n_classes + i] = 0;
+    expected_gini_cnt[2 * n_classes + i] = 0;
+  }
+
+  expected_leaf_cnt[0] = 99;
+  expected_leaf_cnt[1] = 30;
+  expected_leaf_cnt[2] = 71;
+
+  for (int j = 0; j < n_features; ++j) {
+    for (int i = 0; i < n_points; ++i) {
+      if (i < 99) { // leaf 0
+        if (i < 30 || (i >= 100 && i < 130))
+          expected_feature_values[j * n_points + i] =
+              (j & 1) ? ((i != 5 && i != 9) ? 10 : 50) : 50;
+        else
+          expected_feature_values[j * n_points + i] = (j & 1) ? 50 : 10;
+      } else if (i >= 99 && i < 129) { // leaf 1
+        expected_feature_values[j * n_points + i] = (j & 1) ? 10 : 50;
+      } else { // leaf 2
+        expected_feature_values[j * n_points + i] = (j & 1) ? 50 : 10;
+      }
+    }
+  }
+
+  test_check_func(expected_gini_cnt, expected_leaf_cnt, split_feature,
+                  expected_feature_values, 0);
+
+  mem_reset();
+}
+
+#endif
